Seperated show logs into tailer file
